fix(web): restore eager soup/search for authenticated /app
Lazy SearchProvider with an empty default context and lazy SoupView left
inbox, channels, and soup blank. Mount those providers and views eagerly
again, and warm the authenticated workspace graph in Vite.
