+++ apps/web/src/components/app/split-layout/SplitLayoutRoute.tsx
@@ -1,22 +1,14 @@
 import { setGlobalSplitManager } from '@app/signal/splitLayout';
 import type { WithRequired } from '@core/util/withRequired';
 import type { RouteDefinition, RouteSectionProps } from '@solidjs/router';
-import { lazy, Suspense } from 'solid-js';
-
-const SplitLayoutContainer = lazy(() =>
-  import('./SplitLayout').then((module) => ({
-    default: module.SplitLayoutContainer,
-  }))
-);
+import { SplitLayoutContainer } from './SplitLayout';
 
 function LayoutRoute(props: RouteSectionProps) {
   return (
-    <Suspense>
-      <SplitLayoutContainer
-        pairs={props.params.splits?.split('/') ?? []}
-        setManager={setGlobalSplitManager}
-      />
-    </Suspense>
+    <SplitLayoutContainer
+      pairs={props.params.splits?.split('/') ?? []}
+      setManager={setGlobalSplitManager}
+    />
   );
 }
