+++ apps/web/src/components/app/split-layout/componentRegistry.tsx
@@ -1,15 +1,27 @@
 import { useActivityFeedFlag } from '@app/features/activity/use-activity-feed-flag';
 import type { EventEditorInitialValues } from '@app/features/calendar/components/composer/event-form-model';
 import type { CalendarEvent } from '@app/features/calendar/types';
+import { GettingStarted } from '@app/features/getting-started';
+import { Home } from '@app/features/home';
+import { InboxView } from '@app/features/inbox-view/inbox-view';
 import { queryStateFrom } from '@app/features/next-soup/filters/filter-store';
 import type { SetPredicatesInput } from '@app/features/next-soup/filters/filter-store/predicates-store';
 import { mergeQuery } from '@app/features/next-soup/filters/filter-store/query-store';
 import type { Query } from '@app/features/next-soup/filters/filter-store/types';
 import { getViewPreset } from '@app/features/next-soup/sidebar/soup-filter-presets';
+import { NonMemberChannelPreview } from '@app/features/next-soup/soup-view/non-member-channel-preview';
+import { SoupView } from '@app/features/next-soup/soup-view/soup-view';
 import { useRecentViewFlag } from '@app/features/next-soup/use-recent-view-flag';
+import { ReminderEditorSplit } from '@app/features/reminders/ReminderEditorSplit';
+import { SettingsPanelComponentWrapper } from '@app/features/settings/Settings';
 import { useAnalytics } from '@app/lib/analytics/analytics-context';
 import { useFeatureFlag, usePosthog } from '@app/lib/analytics/posthog';
 import { globalSplitManager } from '@app/signal/splitLayout';
+import { EventComposerSplit } from '@block-calendar/components/EventComposerSplit';
+import { ChannelCompose } from '@block-channel/component/Compose';
+import { EmailCompose } from '@block-email/component/compose/Compose';
+import { ComposeSkill } from '@block-md/component/ComposeSkill';
+import { ComposeTask } from '@block-md/component/ComposeTask';
 import {
   CRM_VIEW_URL_PARAM,
   type CrmViewConfig,
@@ -41,34 +53,6 @@ import {
 import type { SplitContent } from './layoutManager';
 import { useSplitPanelOrThrow } from './layoutUtils';
 import { previewEmptyStateForContent } from './previewController';
-
-const GettingStarted = lazy(() =>
-  import('@app/features/getting-started').then((module) => ({
-    default: module.GettingStarted,
-  }))
-);
-const Home = lazy(() =>
-  import('@app/features/home').then((module) => ({
-    default: module.Home,
-  }))
-);
-const InboxView = lazy(() =>
-  import('@app/features/inbox-view/inbox-view').then((module) => ({
-    default: module.InboxView,
-  }))
-);
-const NonMemberChannelPreview = lazy(() =>
-  import('@app/features/next-soup/soup-view/non-member-channel-preview').then(
-    (module) => ({
-      default: module.NonMemberChannelPreview,
-    })
-  )
-);
-const SoupView = lazy(() =>
-  import('@app/features/next-soup/soup-view/soup-view').then((module) => ({
-    default: module.SoupView,
-  }))
-);
 
 function usePageViewTracking(pageTitle: string) {
   const analytics = useAnalytics();
@@ -625,41 +609,6 @@ registerComponent('non-member-channel', (params) => {
     />
   );
 });
-const ChannelCompose = lazy(() =>
-  import('@block-channel/component/Compose').then((module) => ({
-    default: module.ChannelCompose,
-  }))
-);
-const EmailCompose = lazy(() =>
-  import('@block-email/component/compose/Compose').then((module) => ({
-    default: module.EmailCompose,
-  }))
-);
-const ComposeTask = lazy(() =>
-  import('@block-md/component/ComposeTask').then((module) => ({
-    default: module.ComposeTask,
-  }))
-);
-const EventComposerSplit = lazy(() =>
-  import('@block-calendar/components/EventComposerSplit').then((module) => ({
-    default: module.EventComposerSplit,
-  }))
-);
-const ComposeSkill = lazy(() =>
-  import('@block-md/component/ComposeSkill').then((module) => ({
-    default: module.ComposeSkill,
-  }))
-);
-const ReminderEditorSplit = lazy(() =>
-  import('@app/features/reminders/ReminderEditorSplit').then((module) => ({
-    default: module.ReminderEditorSplit,
-  }))
-);
-const SettingsPanelComponentWrapper = lazy(() =>
-  import('@app/features/settings/Settings').then((module) => ({
-    default: module.SettingsPanelComponentWrapper,
-  }))
-);
 
 registerComponent('channel-compose', () => {
   usePageViewTracking('channel-compose');
