+++ apps/web/src/features/next-soup/search-context.tsx
@@ -1,5 +1,5 @@
-export { SearchProvider } from '@app/features/soup/search/SearchProvider';
 export {
+  SearchProvider,
   type SoupSearchPoolItem as SearchPoolItem,
   useSearchContext,
 } from '@app/features/soup/search/context';
