+++ apps/web/src/features/soup/search/index.ts
@@ -1,4 +1,5 @@
 export {
+  SearchProvider,
   type SoupSearchPoolEntry,
   type SoupSearchPoolItem,
   useOptionalSearchContext,
@@ -12,7 +13,6 @@ export {
   type SoupSearchRequest,
   soupSearchMatchType,
 } from './create-search-state';
-export { SearchProvider } from './SearchProvider';
 export {
   createSoupFreshSearch,
   intersectEntityPools,
